+++ src/java/org/apache/cassandra/index/sai/disk/v3/V3VectorPostingsWriter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.vector;
+package org.apache.cassandra.index.sai.disk.v3;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -29,10 +29,11 @@ import java.util.function.IntUnaryOperator;
 
 import io.github.jbellis.jvector.graph.RandomAccessVectorValues;
 import io.github.jbellis.jvector.vector.types.VectorFloat;
+import org.apache.cassandra.index.sai.disk.vector.VectorPostings;
 import org.apache.cassandra.io.util.SequentialWriter;
 import org.apache.cassandra.utils.Pair;
 
-public class VectorPostingsWriter<T>
+public class V3VectorPostingsWriter<T>
 {
     // true if vectors rows are 1:1 (all vectors are associated with exactly 1 row, and each row has a non-null vector)
     private final boolean oneToOne;
@@ -41,7 +42,7 @@ public class VectorPostingsWriter<T>
     // given a "new" ordinal (0..size), return the ordinal it corresponds to in the original graph and VectorValues
     private final IntUnaryOperator newToOldMapper;
 
-    public VectorPostingsWriter(boolean oneToOne, int graphSize, IntUnaryOperator mapper) {
+    public V3VectorPostingsWriter(boolean oneToOne, int graphSize, IntUnaryOperator mapper) {
         this.oneToOne = oneToOne;
         this.graphSize = graphSize;
         this.newToOldMapper = mapper;
